+++ algorithms/kernel/kmeans/inner/kmeans_distributed_v1.h
@@ -1,4 +1,4 @@
-/* file: kmeans_distributed.h */
+/* file: kmeans_distributed_v1.h */
 /*******************************************************************************
 * Copyright 2014-2020 Intel Corporation
 *
@@ -22,13 +22,13 @@
 //--
 */
 
-#ifndef __KMEANS_DISTRIBITED_H__
-#define __KMEANS_DISTRIBITED_H__
+#ifndef __KMEANS_DISTRIBITED_V1_H__
+#define __KMEANS_DISTRIBITED_V1_H__
 
 #include "algorithms/algorithm.h"
 #include "data_management/data/numeric_table.h"
 #include "services/daal_defines.h"
-#include "algorithms/kmeans/kmeans_types.h"
+#include "algorithms/kernel/kmeans/inner/kmeans_types_v1.h"
 
 namespace daal
 {
@@ -36,7 +36,7 @@ namespace algorithms
 {
 namespace kmeans
 {
-namespace interface2
+namespace interface1
 {
 /**
  * @defgroup kmeans_distributed Distributed
@@ -159,7 +159,11 @@ public:
      *  \param[in] nClusters  Number of clusters
      *  \param[in] assignFlag Flag to calculate partial assignment
      */
-    Distributed(size_t nClusters, bool assignFlag = false);
+    Distributed(size_t nClusters, bool assignFlag = false) : parameter(nClusters, 1)
+    {
+        initialize();
+        parameter.assignFlag = assignFlag;
+    }
 
     /**
      * Constructs K-Means algorithm by copying input objects and parameters
@@ -167,7 +171,12 @@ public:
      * \param[in] other An algorithm to be used as the source to initialize the input objects
      *                  and parameters of the algorithm
      */
-    Distributed(const Distributed<step1Local, algorithmFPType, method> & other);
+    Distributed(const Distributed<step1Local, algorithmFPType, method> & other) : parameter(other.parameter)
+    {
+        initialize();
+        input.set(data, other.input.get(data));
+        input.set(inputCentroids, other.input.get(inputCentroids));
+    }
 
     /**
     * Returns the method of the algorithm
@@ -225,18 +234,6 @@ public:
         return services::SharedPtr<Distributed<step1Local, algorithmFPType, method> >(cloneImpl());
     }
 
-    /**
-    * Gets parameter of the algorithm
-    * \return parameter of the algorithm
-    */
-    ParameterType & parameter() { return *static_cast<ParameterType *>(_par); }
-
-    /**
-    * Gets parameter of the algorithm
-    * \return parameter of the algorithm
-    */
-    const ParameterType & parameter() const { return *static_cast<const ParameterType *>(_par); }
-
 protected:
     virtual Distributed<step1Local, algorithmFPType, method> * cloneImpl() const DAAL_C11_OVERRIDE
     {
@@ -265,10 +262,12 @@ protected:
     {
         Analysis<distributed>::_ac = new __DAAL_ALGORITHM_CONTAINER(distributed, DistributedContainer, step1Local, algorithmFPType, method)(&_env);
         _in                        = &input;
+        _par                       = &parameter;
     }
 
 public:
-    InputType input; /*!< %Input data structure */
+    InputType input;         /*!< %Input data structure */
+    ParameterType parameter; /*!< K-Means parameters structure */
 
 private:
     PartialResultPtr _partialResult;
@@ -307,7 +306,11 @@ public:
      *  \param[in] nClusters   Number of clusters
      *  \param[in] nIterations Number of iterations
      */
-    Distributed(size_t nClusters, size_t nIterations = 1);
+    Distributed(size_t nClusters, size_t nIterations = 1) : parameter(nClusters, nIterations)
+    {
+        initialize();
+        parameter.assignFlag = false;
+    }
 
     /**
      * Constructs K-Means algorithm by copying input objects and parameters
@@ -315,7 +318,11 @@ public:
      * \param[in] other An algorithm to be used as the source to initialize the input objects
      *                  and parameters of the algorithm
      */
-    Distributed(const Distributed<step2Master, algorithmFPType, method> & other);
+    Distributed(const Distributed<step2Master, algorithmFPType, method> & other) : parameter(other.parameter)
+    {
+        initialize();
+        input.set(partialResults, other.input.get(partialResults));
+    }
 
     /**
     * Returns the method of the algorithm
@@ -398,18 +405,6 @@ public:
         return services::SharedPtr<Distributed<step2Master, algorithmFPType, method> >(cloneImpl());
     }
 
-    /**
-    * Gets parameter of the algorithm
-    * \return parameter of the algorithm
-    */
-    ParameterType & parameter() { return *static_cast<ParameterType *>(_par); }
-
-    /**
-    * Gets parameter of the algorithm
-    * \return parameter of the algorithm
-    */
-    const ParameterType & parameter() const { return *static_cast<const ParameterType *>(_par); }
-
 protected:
     virtual Distributed<step2Master, algorithmFPType, method> * cloneImpl() const DAAL_C11_OVERRIDE
     {
@@ -438,10 +433,12 @@ protected:
     {
         Analysis<distributed>::_ac = new __DAAL_ALGORITHM_CONTAINER(distributed, DistributedContainer, step2Master, algorithmFPType, method)(&_env);
         _in                        = &input;
+        _par                       = &parameter;
     }
 
 public:
-    InputType input; /*!< %Input data structure */
+    InputType input;         /*!< %Input data structure */
+    ParameterType parameter; /*!< K-Means parameters structure */
 
 private:
     PartialResultPtr _partialResult;
@@ -450,10 +447,7 @@ private:
     Distributed & operator=(const Distributed &);
 };
 /** @} */
-} // namespace interface2
-
-using interface2::DistributedContainer;
-using interface2::Distributed;
+} // namespace interface1
 } // namespace kmeans
 } // namespace algorithms
 } // namespace daal
