+++ examples/cpp/source/kmeans/kmeans_init_csr_distr.cpp
@@ -249,7 +249,7 @@ void calculateCentroids(const NumericTablePtr & initialCentroids, const NumericT
     }
 
     /* Print the clusterization results */
-    printNumericTable(assignments[0], "First 10 cluster assignments from 1st node:", 10);
     printNumericTable(centroids, "First 10 dimensions of centroids:", 20, 10);
+    printNumericTable(assignments[0], "First 10 cluster assignments from 1st node:", 10);
     printNumericTable(objectiveFunction, "Objective function value:");
 }
